pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
]
